Time: 4 ms (79.63%), Space: 9.6 MB (53.96%) - LeetHub

+++ 0074-search-a-2d-matrix/0074-search-a-2d-matrix.cpp
@@ -1,33 +1,25 @@
 class Solution {
 public:
     bool searchMatrix(vector<vector<int>>& matrix, int target) {
-        long long hor = matrix.size();
-        long long ver = matrix[0].size();
-        //cout << hor << ver << endl;
-        long long left = 0, right = ver -1 ;
-        long long root = matrix[left][right];
+        int hor = matrix.size();
+        int long ver = matrix[0].size();
+        int left = 0, right = ver -1 ;
+        int root = matrix[left][right];
         while(left < hor && right < ver){
             if(root == target){
-                //cout << root << endl;
                 return true;
             }
             else if(root > target){
                 if(right == 0)
                     break;
-                //cout << "root is bigger : left, right" << left << ',' << right << ' ' << root << endl;
                 right -= 1;
                 root = matrix[left][right];
-                
-                //cout << "left, right" << left << ',' << right << ' ' << root << endl;
             }
             else{
                 if(left == (hor-1))
                     break;
-               // cout << "root is smaller : left, right" << left << ',' << right << ' ' << root << endl;
                 left += 1;
                 root = matrix[left][right];
-               // cout << "left, right" << left << ',' << right << ' ' << root << endl;
-
             }
         }
         return false;
